Changes duration base value

+++ athletictraining/src/main/kotlin/it/unibo/lss/fcla/athletictraining/domain/model/exercise/Duration.kt
@@ -8,18 +8,18 @@ import it.unibo.lss.fcla.athletictraining.domain.model.exercise.exception.Durati
  *
  * @author Nicola Lasagni on 07/04/2021.
  */
-class Duration(val seconds: Int) {
+class Duration(val seconds: Long) {
 
     companion object {
 
         private const val MINUTES_TO_SECONDS = 60
         private const val HOURS_TO_SECONDS = MINUTES_TO_SECONDS * 60
 
-        fun ofMinutes(minutes: Int): Duration {
+        fun ofMinutes(minutes: Long): Duration {
             return Duration(minutes * MINUTES_TO_SECONDS)
         }
 
-        fun ofHours(hours: Int): Duration {
+        fun ofHours(hours: Long): Duration {
             return Duration(hours * HOURS_TO_SECONDS)
         }
     }
